+++ ServerManagementTools/Public/Get-IISLogPath.ps1
@@ -4,8 +4,6 @@ function Get-IISLogPath {
         Retrieve webiste logging path.
     .DESCRIPTION
         The Get-IISLogPath cmdlet retrieves the log file path for one or more websites configured on the target computer.
-    .PARAMETER Name
-        Specifies a name of one or more websites.  Get-IISLogPath retrieves the logging path for the website specified.  If you do not specify this parameter, the cmdlet will return all configured sites.
     .EXAMPLE
         Get-IISLogPath
         Returns log path information for all sites
@@ -24,36 +22,37 @@ function Get-IISLogPath {
         Author: Trent Willingham
         Check out my other scripts and projects @ https://github.com/twillin912
     #>
-
     [CmdletBinding()]
-    [OutputType([PSObject])]
-    Param (
-        [Parameter(ValueFromPipeline=$true)]
-        [string[]] $Name
+    [OutputType([psobject])]
+    param(
+        # Specifies a name of one or more websites.  Get-IISLogPath retrieves the logging path for the website specified.  If you do not specify this parameter, the cmdlet will return all configured sites.
+        [parameter(ValueFromPipeline = $true)]
+        [string[]]$Name
     )
 
-    Begin {
+    begin {
         $WebsiteObjects = Get-Website
         $FilteredSites = @()
     }
 
-    Process {
-        if ( $Name ) {
-            foreach ( $SiteName in $Name ) {
+    process {
+        if ($Name) {
+            foreach ($SiteName in $Name) {
                 $FilteredSites += $WebsiteObjects | Where-Object { $PSItem.Name -like $Sitename }
             }
-        } else {
+        }
+        else {
             $FilteredSites = $WebsiteObjects
         }
 
-        foreach ( $Site in $FilteredSites ) {
+        foreach ($Site in $FilteredSites) {
             $LogPath = "$($Site.logFile.directory)\W3SVC$($Site.id)"
             $LogPath = [System.Environment]::ExpandEnvironmentVariables($LogPath)
 
             $Object = New-Object -TypeName PSCustomObject -Property @{
-                Id          = $Site.id
-                Name        = $Site.name
-                LogPath     = $LogPath
+                Id      = $Site.Id
+                Name    = $Site.Name
+                LogPath = $LogPath
             }
             $Object.PSObject.TypeNames.Insert(0, 'ServerManagementTools.IISLogPath')
             Write-Output -InputObject $Object