+++ module.settings.ps1
@@ -7,7 +7,7 @@ $Author         = 'Trent Willingham'
 [System.Diagnostics.CodeAnalysis.SuppressMessage('PSUseDeclaredVarsMoreThanAssigments', '')]
 $Description    = 'Collection of PowerShell functions for managing Windows Server.'
 [System.Diagnostics.CodeAnalysis.SuppressMessage('PSUseDeclaredVarsMoreThanAssigments', '')]
-$Tags           = @()
+$Tags           = @('IIS','Log','Logs','RDP','Rotate','Rotation','SChannel','Security')
 
 [System.Diagnostics.CodeAnalysis.SuppressMessage('PSUseDeclaredVarsMoreThanAssigments', '')]
 $ProjectUri     = "https://github.com/twillin912/${env:BHProjectName}"
@@ -20,4 +20,3 @@ $ReleaseNotes   = "${ProjectUri}/blob/stable/RELEASENOTES.md"
 
 [System.Diagnostics.CodeAnalysis.SuppressMessage('PSUseDeclaredVarsMoreThanAssigments', '')]
 $GitRepo        = "twillin912/${env:BHProjectName}"
-