+++ CHANGELOG.md
@@ -1,23 +1,40 @@
 # ServerManagementTools Release History
 
-## 0.4.0 (2017-04-05)
+## v0.5 (2018-04-24)
 
 - New Functions
-  - `Get-RDSession`: This function uses the Cassia.dll to query local or remote computers for active and disconnection Remote Desktop sessions.
+
+  - The `Get-LinuxCdpInfo` cmdlet queries a Linux server for CDP information by capturing the network packets using tcpdump.
+  - The `Install-DiskCleanupTool` cmdlet copies the Disk Cleanup executable and supporting files from the WinSxS folder the to correct installed location and creates the shortcut.
+
 - Updates
+
+  - Invoke-LogRotation: Refactored the parameter name 'CompressDays' to 'KeepRaw'
+  - Register-LogRotationTask: Refactored the parameter name 'CompressDays' to 'KeepRaw'
+
+
+## v0.4 (2017-04-05)
+
+- New Functions
+
+  - The `Disable-SChannelFeature` cmdlet disables features in the SChannel security suite on Windows computers.  This cmdlet can be used to disable ciphers, key exchanges, and protocols that are consider insecure.
+  - The `Get-RDSession` cmdlet uses the Cassia.dll to query local or remote computers for active and disconnection Remote Desktop sessions.
+
+- Updates
+
   - Updated the build process use utilize the InvokeBuild module.
 
-## 0.3.0 (2017-02-17)
+## v0.3 (2017-02-17)
 
 - New Functions
   - `Get-IISLogPath`: This function uses the WebAdministration module to query the IIS configuration and retreive the log file locations.
 
-## 0.2.0 (2017-02-16)
+## v0.2 (2017-02-16)
 
 - New Functions
   - `Invoke-LogRotation`: Compresses log files by month
 
-## 0.1.0 (2017-02-15)
+## v0.1 (2017-02-15)
 
 - New Functions
   - `Get-DfsrBacklogStatus`: Query DFSR replication backlog