+++ ReleaseNotes.md
@@ -1,9 +1,12 @@
-# What is new in ServerManagementTools 0.4.0
+# What is new in ServerManagementTools v0.5
 
 ## New Functions
 
-- `Get-RDSession`: This function uses the Cassia.dll to query local or remote computers for active and disconnection Remote Desktop sessions.
+- The `Get-LinuxCdpInfo` cmdlet queries a Linux server for CDP information by capturing the network packets using tcpdump.
+
+- The `Install-DiskCleanupTool` cmdlet copies the Disk Cleanup executable and supporting files from the WinSxS folder the to correct installed location and creates the shortcut.
 
 ## Updates
 
-- Updated the build process use utilize the InvokeBuild module.
+- Invoke-LogRotation: Refactored the parameter name 'CompressDays' to 'KeepRaw'
+- Register-LogRotationTask: Refactored the parameter name 'CompressDays' to 'KeepRaw'